Move operators into their own file
