+++ pyccel/codegen/printing/cwrappercode.py
@@ -12,9 +12,11 @@ from pyccel.ast.literals  import LiteralTrue, LiteralInteger
 from pyccel.ast.builtins import PythonPrint
 
 from pyccel.ast.core import Variable, ValuedVariable, Assign, AliasAssign, FunctionDef, FunctionAddress
-from pyccel.ast.core import If, Nil, Return, FunctionCall, PyccelNot
+from pyccel.ast.core import If, Nil, Return, FunctionCall
 from pyccel.ast.core import create_incremented_string, SeparatorComment
-from pyccel.ast.core import VariableAddress, Import, PyccelNe
+from pyccel.ast.core import VariableAddress, Import
+
+from pyccel.ast.operators import PyccelNe, PyccelNot
 
 from pyccel.ast.datatypes import NativeInteger, NativeBool, NativeComplex, NativeReal
